chat-ui/requirements.txt
fastapi==0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
mcp[cli]>=1.20.0
openai>=1.50.0
anthropic>=0.39.0
google-genai>=1.0.0
